requirements.txt
zxcvbn==4.4.28 
nltk==3.8.1 
requests==2.31.0 
python-dotenv==1.0.0 
flask==3.0.2 